Fixed an Issue that caused Message Drops
It was not correctly checked wether we really received a new address value from ZMQ, which induced a potential and temporary swap of address and message fields. This could yeald to us parsing nonsense. This, in turn, led to us throwing away many messages and thus imposed a huge performance problem.